Unconditionally set roProp bit to elide earlier stores
Summary:
Prop-Dim/BaseSC/BaseL previously only set the `roProp` field of the MInstrState when they weren't going to Throw. Since we can now Warn, the generated IR after these instructions can have a phi from the branch where the `roProp` bit was set to True, and the branch where we `ThrowOrWarn` and didn't set it.

We should just set it each time we encounter a `CheckROCOW` or `readonly + CheckMutROCOW`, so we can always elide the `StMROProp` that the BaseX instructions emit and get rid of this phi.

**[RFC]** In addition, change defintion of "COW" to be is not KindOfObject. We only care about things that can be indexed into with an Elem-Dim that are not copy-on-write. Clsmeths no longer promote to vecs, so I believe we only care about collections.

Reviewed By: ricklavoie

Differential Revision: D31215937

fbshipit-source-id: e66665ea85e2b35aacfb238917db67e048c5728a