Check Dim for Readonly
Summary:
```
public static readonly vec<Map<arraykey, int>> $c3 = vec[Map {'a' => 5}];
Foo::$c3[0]['a'] = 5;
```

Readonly COW types can be modified. To enforce this, BaseSC/BaseL/Prop-Dims can set a bit in the MInstrState, roProp, that tells every sequential Elem-Dim it needs to check whether its COW or not. If not, it will need to throw.

This was implemented with a suffix to `emitDim` and `emitBespokeDim`

Note:  My definition of COW in this case is not Obj type. I believe the only non-value types that can occur in these positions where we are indexing into them are Collections.
[RFC] Are there other types I should include in this defintion? This is also pretty easy to change.

Reviewed By: ricklavoie

Differential Revision: D30948085

fbshipit-source-id: 33017b2d9431dcc45086d3825ef4a76a5d201b61