Introduce supportdynamic type hint and Tsupportdynamic
Summary:
This introduces all the boilerplate for a new hint `Hsupportdynamic` and type `Tsupportdynamic`.

We use the standard naming transformation `Happly ("supportdynamic", []) => Hsupportdynamic` as used for nonnull, dynamic, etc. as opposed to mucking with the parser.

I've left subtyping stubbed out in this diff; this is the sole focus of {D31137209}.

**Notes**
Since the type is just for pessimization at the moment and hasn't gone through design review, I've used the simplest possible gating and skipped building runtime support entirely. A file with supportdynamic will just fatal on an unknown type at runtime.

We will probably want `supportdynamic` to enforce as `nonnull` and `?supportdynamic` to enforce as `mixed` i.e. not enforce.

Differential Revision: D31061074

fbshipit-source-id: 39de015981d35b046636112560bdc20ac7e12b18