Avoid the need to stub out JK in OSS builds
Summary:
In {D29479254 (https://github.com/facebook/hhvm/commit/8cb5727813b937f20ffe180662df9cbbe82438f4)}, we added some fishy code which returns a specific error string from `eval` and `get` for OSS builds, and checks for that string in ServerLocalConfig. We did this because the JustKnobs interface is an FB-internal one which shouldn't be polluted with hh-specific logic.

Instead of using the JustKnobs interface directly (and stubbing it out for the OSS build), stub out `apply_justknobs_overrides`, so that we need not worry about how the JK interface should behave in OSS.

Also, instead of logging an error when lookups fail, raise an exception (as recommended by joshkehn on D31213720).

Differential Revision: D31246563

fbshipit-source-id: 4578888535d5b0925b2d3a09095e2aaef8882d71