Unpublished frame tracking in backtraces
Summary: Link the unpublished frame offset tracking with backtrace routines. The current implementation only exposes this for the leafmost frame, which is sufficient for coeffects. A follow-up diff will clean up this API and also pass the unpublished frames for parent frames, enabling us to elide frames when metadata is present.

Reviewed By: oulgen

Differential Revision: D31119297

fbshipit-source-id: cb4a27fbb190a6cd6728414a5d6797475b4de08c