Adding more comments.
Added comments in `agent_graph_generator.lp` to describe the main idea of this logic program. Also how the `depends_on` edges are been selected by combining with agent_connector function in the Python interface.

Added comments in `agentGraphGenerator.py` to illustrate the meaning of each command line option as well as the connector function.