Remove unused and unusable KillOnTimeout option
Summary: While searching for instances of CR_NET_READ_INTERRUPTED and CR_NET_WRITE_INTERRUPTED, I found this section of code in the MySQL HHVM extension that is designed to send a KILL command to the server when a client times out while doing a query.  It won't ever get called as expected because the error numbers are wrong.  It probably won't work anyway if we fix the error messages because the code looks like it is way out of date (who uses passwords anymore).  And, it is not currently enabled anyway.  Just rip it out.

Reviewed By: abal147

Differential Revision: D31156976

fbshipit-source-id: 8767d4ab6c25b4579cfe5903281be984b86dec13