Use versioning to avoid querying the DB during updates
Summary:
Native Facts stores facts in two places.

* The in-memory map contains facts which have been inserted since the HHVM process started. All data in the in-memory map is guaranteed to be current, but the in-memory map may be incomplete.
* The DB contains facts representing some point in the past. This data may be days old, and may have been downloaded from a different machine. The DB is guaranteed to be complete, but it may not represent the repo as it currently is.

When we erase a path, we need to ensure that we never return information that was stored in a path that is now erased.

Before this diff, we solved this problem by querying the DB for old information, erasing the relevant information from the data we queried, then storing it in our in-memory map. Then, subsequent queries can just fetch the information from our in-memory map.

But this meant we were querying the DB during a time at which no code can run. We're also querying data which the user may never actually want.

This diff removes those DB queries from `SymbolMap::removePath()`, simplifying the function so it just bumps the path's version number. At read time, we'll filter out the information that involves stale version numbers.

Reviewed By: bdchik

Differential Revision: D30853663

fbshipit-source-id: 46d49bd72ca81ec4191493ba52db6ac0d6b4aeab