Add target for rust ffi, reduce parrellism of FB ubuntu CI jobs
Summary:
Now that cmake understands which parts of HHVM depend on which rust libraries - and how to build them -
it parallelizes them more; this is good, but increases resource usage.

I accidentally reproduced the most recent issue from T100322687, and it turned out to be an OOM.

Without this diff, we end up with all of these at the same time:
- 3 cargo processes
- up to NCORES rustc processes per cargo process
- up to (NCORES-3) other build steps (e.g. dune, g++)

... for a grand total of ((4*ncores)-3) processes fighting for memory. Pre-building one part
sequentially brings us down to a max of ncores processes, which shouldn't affect build time, but should reduce peak memory usage.

Reviewed By: kavoor

Differential Revision: D31319519

fbshipit-source-id: 55de35b3cf55e5a9b140a284cfcb130b171dacf4