Add PairSmartConstructors
Summary:
The goal of this stack is to provide a parser which produces both a concrete syntax tree and Decls (the output of the direct decl parser) in a single parsing pass.

We can get the FFP to construct both trees at once using a SmartConstructors implementation that defers to two other SmartConstructors and pairs their outputs. This diff adds such a SmartConstructors implementation.

Reviewed By: shayne-fletcher

Differential Revision: D30192691

fbshipit-source-id: 3cdd822e4d43d234899f3ad2940638ec1058c73c