Agent graph generation (#8902)
Summary:
A Python binary that takes a set of parameters about the agent graph to produce a valid agent graph for the next phase.

The metrics included so far:
1. number of agents.
2. number of roots.
3. number of leaves.
4. number of infra agents.
5. number of product agents.
6. number of levels and the agent distribution at each level.
7. infra agent profile.
8. product agent profile.

Metric 1, 2, 6 specified in an array of integers as the input.
[2, 4, 4] means we have 10 agents in total, 2 of these are roots. And these agents are distributed in three levels with number of 2, 4, 4 agents at each level accordingly.

The number of leaves in this case is 4 at least. The user can specify a minimum number other than 4. For example, if we specify the leaves to be 5, which means there is one agent in the second level didn't connect to the third level. The sum of infra agents and product agents couldn't exceed the number of total agents.

The profile for each type of the agent, specified how many other agents could related with. For example, infra agent in degree [2, 10] means an infra agent should at least depend on other 2 agents, and couldn't exceed 10.

A set of customizable boundary functions as well as a set of customizable connector functions are provided for easy experimenting on different scenarios in the future.

Pull Request resolved: https://github.com/facebook/hhvm/pull/8902

Test Plan:
Under `GeneratingAgentDistributionTest`, three test cases covered the single level, multiple level, and invalid level inputs.

Under `GeneratingAnAgentGraphTest`, one test case covered the usage of customizable function/parameters to get variety results, the other one test case covered the "unsatisfiable" case.

Reviewed By: hgoldstein, shayne-fletcher

Differential Revision: D31294599

Pulled By: jamesjwu

fbshipit-source-id: 730047324b172f217311a10e3ec17ec38bdf17ee