Introduce readonly variant of idx
Summary:
This diff introduces idxReadonly, which acts as idx on a readonly collection (and returns a readonly result).

Because we do not have ways to parameterize on readonly (nor do we intend to build such a feature in the near future), some collection HSL functions will need readonly variants, like idx.

Reviewed By: oulgen

Differential Revision: D31317828

fbshipit-source-id: 363edba72d737abf227950870c33987414e92857