Stabilise test output
Summary:
The test output was sensitive to the order constraints are inserted. We now sort the output to make it more resistant to irrelevant changes.

Of course, a better solution is to use a set of constraints, but sadly static key constraint has an expression and a type in it which prevents me from deriving `ord` that is necessary to create a set of constraints. The solution out of this (without defining an ordering on the AST and the types) is to have two maps one for expressions and one for types indexed by integers and use those integers in the relevant constraints. However, this is a bit too much bookkeeping for the moment, so I'm doing the easy solution of sorting the output.

Reviewed By: enetsee

Differential Revision: D30541684

fbshipit-source-id: ebe680019207c70c33088261add466db5fa6789c