Implement "submodules"
Summary:
In the current module prototype, there is no central repository for module info: any information needs to be derived from a single declaration. This is generally true for Hack, but the other decls have ways of traversing the declaration tree for getting this info. For example, the following:
```
function upcast(Foo $f): IBar { return $f; }
```
The subtyping check here will traverse the hierarchy of `Foo` until it can show that `Foo <: IBar`.

Without futzing *too* much with how we represent modules, let's try to establish some module relationships. Using the names of modules, fully qualified, we establish "submodules." The new rule added is that submodules are allowed to access the internal members of their parents but *not* vice-versa.

Reviewed By: Wilfred

Differential Revision: D31251524

fbshipit-source-id: 7908dd01f7f4a94e9a564d8265fd2b71ef877e70