Int.compare in ifiles-replacement
Summary:
We used to use `let compare = ( - )`. But if this is asked to compare -4368257400954407633 against 4576980744340540367 (say) the result of the subtraction won't fit inside an ocaml signed int, so it will "underflow" and wrap around to a positive integer, claiming therefore that the negative number is larger than the positive number!

This diff changes only the places relating to Typing_deps.Dep.t. I wanted to judge perf impact solely of these things. And I think that Dep.t is more likely to have large positives and negatives than other integers, perhaps!

Reviewed By: bobrenjc93

Differential Revision: D31433453

fbshipit-source-id: 2e3a6b912fa26c6f04ffe18ce920492d5d01adf5