Run readonly check based on presence of readonly keyword instead of file attribute
Summary:
This modifies the typechecker readonly check so that it only runs when there's a readonly keyword in the function body. (Note that the parser local checks run in the presence of any readonly keyword in the function definition as well, and adds readonly expressions as necessary).

This allows us to separate the turning on/off of readonly check from the preview feature (thus allowing us to roll out the feature without any impact on perf).

This does *not* change the behavior of requiring the UnstableFeature for HackC or the parser yet, we'll do that once we're closer to the launch date.

Reviewed By: oulgen

Differential Revision: D31374593

fbshipit-source-id: 28214fd7076766c4abed24ac1165ad1a3143a5a5