Error for parent::__construct in traits when parent is not ConsistentConstruct
Summary:
Traits allow calls to `parent::__construct` in an unsafe way.

```
<?hh

class Top {
  public function __construct(mixed $m): void {}
}

class Middle extends Top {
  public function __construct(int $i): void {
    parent::__construct($i);
  }
}

trait Badness {
  require extends Top;
  public function __construct(mixed $m): void {
    parent::__construct($m);
  }
}

class Bottom extends Middle {
  use Badness;
}

<<__EntryPoint>>
function main(): void {
  new Bottom("hello"); // exception
}
```
In this case, the trait thinks it's calling `Top::__construct` via its require extends clause, but  instead `Middle::__construct` is called, so the string `"hello"` causes a type hint violation exception.

This diff adds an extra requirement to a `parent::__construct` call in a trait: the parent that provides the constructor must be <<__ConsistentConstruct>>. Making this change would raise a Hack error on the definition of `Middle::__construct` in the example above.

**Note:** This change is morally equivalent to treating `parent::__construct` in a trait like
```
$x = parent::class;
new $x();
```
which currently raises an error if the parent is not <<__ConsistentConstruct>>.

Reviewed By: jamesjwu

Differential Revision: D31359162

fbshipit-source-id: aba4987ae7c7b769a5ea3cb5888f1a1180ab6f6d