Fix some array index bugs
Summary:
WRT sound dynamic, previously, array index operations on `dynamic` didn't check that the value assigned was a subtype of `dynamic`. This check is necessary to make sure that only things that support dynamic can be set into a dynamic with indexing.

Also, there was a soundness bug with both sound dynamic and previous dynamic both as follows.

$dict : dict<int, int>
$d : dynamic
$dict[$d] = 1
$dict : dict<dynamic | int, int>

But we know that the assignment can only succeed if the dynamic is actually an arraykey, so in that case, just make the resulting $dict have type dict<arraykey, int>. Otherwise the `dict` has an inconsistent type that violates its `as arraykey` constraint. This makes it a `nothing` factory in code that is in fact reachable.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D31308375

fbshipit-source-id: 2e67814fd287fbdcca54a76a5b94606abd618590