Int.compare not ( - )
Summary:
We use `let compare = ( - )` in a lot of places. But if this is asked to compare -4368257400954407633 against 4576980744340540367 (say) the result of the subtraction won't fit inside an ocaml signed int, so it will "underflow" and wrap around to a positive integer, claiming therefore that the negative number is larger than the positive number!

The previous diff in the stack fixed a case where I'd observed this to be causing problems. Are there any other cases where it's causing problems? I don't know. For this diff I merely did a BigGrep search for everywhere that uses "let compare" for integers with subtraction, and changed it to Int.compare.

Reviewed By: fahndrich, jamesjwu

Differential Revision: D31431837

fbshipit-source-id: 5e28f78499a4d2f3b7f99e46916ae63a2eba01ed