Simplify remaining COW type definition
Summary:
D31215937 (https://github.com/facebook/hhvm/commit/833005e822e48844331fbe6c9be61e7b47838380) changes definition of a "COW" type to be "non-Obj" for readonly property enforcement because the only legal type that is not COW that could appear in a member-op sequence that gets indexed is a Obj.

Clsmeths no longer promote to vecs so checking if something is an Obj is an equivalent simpler check than what we do here.

Reviewed By: ricklavoie

Differential Revision: D31493794

fbshipit-source-id: 2143c0bdb0e780b4ab78ed618180dc9320978643