Make resumableStackBase() work for first FPs in outer VMs
Summary:
resumableStackBase() assumed that it is used only for the first FP in the
current VM instance, but the new backtrace logic will use it also for FPs
resumed by the asio scheduler in other VMs.

Let's use getPrevVMState() to look up the previous SP, which is just a
generalized version of the current code.

Reviewed By: ricklavoie

Differential Revision: D31528573

fbshipit-source-id: 7bbb998dbec3cf30defa7e9a7b2140e68bdac766