Back BTFrames with IFrames instead of fake ActRecs
Summary:
Instead of keeping track of two ActRecs in BTFrame and storing extra metadata
in BTContext, we just store IFrameID of the current and published frame in
BTFrame and kill most of the BTContext.

BTFrame accessors check if we are in an inlined frame and apply a custom logic.

The depth of IFrame chain is no longer tracked by a counter, instead we store
the published frame in addition to the one being processed. This is useful for
O(1) calculation of unspilled ActRec position at runtime, as we move from
published stack base to the current stack base via root stack base.

This diff makes it possible to access locals of inlined frames.

Reviewed By: ricklavoie

Differential Revision: D31526763

fbshipit-source-id: 3db88cdb9ff6ea9ad5b6c803df0dfa98423c11fd