Move Autoload.ExcludedRepos to RuntimeOption to avoid the lint
Summary:
HHVM emits a warning when it receives a runtime option that it doesn't use:

```
Possible bad config node: Autoload.ExcludedRepos
```

Facts only checks the `Autoload.ExcludedRepos` option when `Autoload.Enabled` is true, so people who aren't currently using native Facts are seeing this.

The fix is pretty simple - we just add the option to the `RuntimeOption` class. Then it will always be used.

Reviewed By: kavoor

Differential Revision: D31522463

fbshipit-source-id: 26815e312391656d5aa0303336a6dbefc65338c3