Cache decls in full inits
Summary: In the previous diff, we opted not to cache decls in Direct_decl_service. This ensures that in saved state inits and incremental typechecks, we don't overwrite old decls during parsing. We can generally afford to do this because the number of files we need to parse is not large, but this isn't the case for full inits! Since there aren't old decls to overwrite in full init scenarios, we can mitigate the perf regression of the previous diff by caching decls during the parsing step.

Reviewed By: bobrenjc93

Differential Revision: D31449725

fbshipit-source-id: b5fa9c8feaa30192e9585eb88adc859850f0e18d