Recheck descendents for shallow decl fanout
Summary: We are currently missing a case where if we add a member to some ancestor class and it's incompatible with an override in a descendent class. This diff makes sure that whenever we change an ancestor class, we add all descendent classes so that we don't miss this case.

Differential Revision: D31556834

fbshipit-source-id: e384894c9c7aad9add8df0b561f3dd461df89ec9