make --pessimise-builtins pessimise collections and arrays
Summary: Pessimise vec, Vector, dict, and Map when the `--pessimise-builtins` option is set. This means that a `vec<T>` is treated as holding either `T` or `dynamic` (`~T`). This just alters the creation and array indexing syntaxes which are implemented directly in the type checker. The types of other functions on them are specified in .hhi files, and not altered here.

Reviewed By: andrewjkennedy

Differential Revision: D31473889

fbshipit-source-id: e2b73cfca3e26ea120c32d78fd012f5c00f81845