Stop committing mcrouter codegen
Now that we're using thrift's build system instead of a custom one,
we have a working `thrift1`, so don't need to commit the codegen.

I'm not doing the usual submodule -> externalproject_add thing here as mcrouter
- does not use CMake
- does not publish regular releases

So, this is still going to be more work to update than most dependencies, but at
least we're no longer going to need to update the codegen by hand.

Test plan: lego*, manually built on AWS (not on my mac as mcrouter does not build on MacOS)