add flag for loading from hack/64 naming table
Summary:
Out of an abundance of caution, ljw1004 suggested adding this flag before rolling out the sandcastle naming table generation. Currently sandcastle jobs by default use the unbacked (non sqlite) naming table. This means we will end up writing 7m rows even for incremental saved state generation situations. This takes 2 minutes, so it should be ok. But to be safe...

This diff adds a flag so if a sqlite naming table already exists in the hack/64 saved state, load from that. This way when we do our incremental sqlite naming table generation, we only need to write the rows that change.

In the next diff I'll update the sandcastle job (that's guarded by a feature) to use this flag.

Differential Revision: D31566176

fbshipit-source-id: 9b57d2f1416e8b9e84dcf172c48c76a3c4a03ee1