Allow Expression Trees in pipe (Excluding $$ usage)
Summary:
Currently, when Expression Trees are used in conjunction with the pipe operator, there are problems:

First, there is the issue of `Other[0000] Feature not implemented: Assignment within pipe expressions`.

We get this error even if Expression Trees do not use the $$ within a splice expression. So for instance:

```
1 |> foo(Code`${ bar() }`, $$)
```

This will produce an error because of the way that we typecheck Expression Trees. We pretend to typecheck the splices (and function pointers) in an expression tree as if those assignments were not in the scope of a lambda. So we pretend to type the following as:

```
1 |> $0tmp0 = bar() in Code::makeTree(...) ...
```

Which will trigger the Assignment within pipe expressions error. This is because the way that we determine whether an assignment occurs within a pipe is by checking for the $$ in scope of the rhs of an assignment. We remove this check because assignments are no longer expressions and therefore cannot occur on the rhs of a pipe.

Furthermore, we now need to implement the capability in Expression Trees to use $$ in splice now that this is no longer a typechecker error but still a runtime error. We do this via a code transformation.

The code transformation principle is as follows, if there is an occurrence of $$ that refers to outside the Expression Tree, we will turn all of those references into the temporary variable `$0dollardollar`, then we will transform the "immediately invoked lambda" that is created when desugaring expression trees into one that takes a new argument `$0dollardollar` and is called with the $$ variable. In short, the following occurs:

```
1 |> Code`${ $$ }`
```

Instead of the following transformation:

```
1 |>
  (() ==> {
    $0splice0 = $$;
    return Code::makeTree(
      ...
      function($0v) use ($0splice0) {
        $0v->splice( $0splice0 )
      },
    );
  })()
```

Which doesn't work because lambdas do not capture $$, we will do the following transformation

```
1 |>
  (($0dollardollar) ==> {
    $0splice0 = $0dollardollar;
    return Code::makeTree(
      ...
      function($0v) use ($0splice0) {
        $0v->splice( $0splice0 )
      },
    );
  })($$)
```

Which doesn't encounter the same problem because the $$ isn't being referred to within the lambda, instead it is being passed into the immediately invoked lambda, so capture by a different name.

When we traverse the splices to replace the existing $$ variables with their new stand in the `$0dollardollar` variable, we need to ensure that we don't capture every reference to `$$`. We should only capture the ones that reference the value on the lhs of the pipe. So for instance:

```
1 |> Code`${ $$ |> foo($$) }`
```

Only the first $$ here should be transformed. Similarly, the $$ variables inside a lambda in a splice should not be transformed

```
1 |> Code`() ==> { $x = 1 |> foo($$); }`
```

This $$ should not be transformed either to retain correctness.

This code transformation is how we make the code work in the codegen path. In the typing path, we need to do something else due to the way that we have hacked typing expression trees.

In typing, we currently do not see the existence of the outer lambda. We type the assignments to temporary variables `$0splice0` as if it were in the same environment as the surrounding code. This is normally fine because of the fact that we immediately invoke the lambda, so we get the benefit of retaining of fake member typing. However, since typing doesn't see the existence of the wrapper lambda, it is hard to get typing to see the existence of a new variable in the environment (`$0dollardollar`). So we do the following, we simply tell typing.ml that such a variable exists and has the same type as `$$`. This way, when typing sees any references to `$0dollardollar`, it ill have the correct typing information.

Reviewed By: Wilfred

Differential Revision: D31253391

fbshipit-source-id: fb23c6a7234407c36d71a8c998d1e0187c6158f2