Convert hhas_to_string entrypoint to cxx
Summary:
Design considerations:

`hackc_hhas_program_raw()` takes our `HhasProgramWrapper` and returns the address of our `HhasProgram` inside. I made this a helper so we can otherwise safely pass around the `HhasProgramWrapper` type. Note that this function does not consume the `HhasProgramWrapper`, which is intentional because we don't want to perform any copies. We would rather just return a pointer to the place in memory in the struct where our `HhasProgram` is. C++ then frees the `HhasProgramWrapper` when it goes out of scope.

Reviewed By: shayne-fletcher

Differential Revision: D31591294

fbshipit-source-id: df9bbf3cf3f24582082af812a46c70d21c8e7d5e