Convert hhas program entrypoint to cxx 1/2
Summary:
This diff just combines `rust_decl_ffi_bridge` with `rust_compile_ffi_bridge`
----
Migrating the `HhasProgram` and `hhas_to_string` entrypoints to cxx

*Issues: [RFC]*
Both the Decl entrypoint and this new HhasProgram entrypoint need the Bump type in their signature. In order to expose this opaque type to C++, we add `type Bump` to the "extern Rust" clause of our cxx interface. If we try to use the same *defintion* by adding a `use bumpalo::Bump`, cxx complains that we are implementing a trait for a type not defined in this crate. Its true, Bump is defined in bumpalo, and I'm guessing this trait we are implementing is part of the cxx magic. In order to get around this, we can simply define a tuple struct `struct Bump(bumpalo::Bump);` in this crate.

```
stderr: error[E0117]: only traits defined in the current crate can be implemented for arbitrary types
  --> hphp/hack/src/facts/ffi_bridge/rust_facts_ffi_bridge.rs:75:9
   |
75 |         type Bump;
   |         ^^^^^----
   |         |    |
   |         |    `Bump` is not defined in the current crate
   |         impl doesn't use only types from inside the current crate
   |
   = note: define and implement a trait or new type instead
```

But, if we do this for both the Decl and HhasProgram entrypoint, we get duplicate symbol errors for functions cxx generates in C++ for this opaque type, which also is true.
```
ld.lld: error: duplicate symbol: cxxbridge1$Bump$operator$sizeof
```

The solution here is just to combine these cxx interfaces into one that can both use the same `Bump` and `NativeEnv` definitions. We also explored reinterpret casting different tuple structs that share the same layout in memory `BumpCompile(bumpalo::Bump)` and ` BumpDecl(bumpalo::Bump)`, but decided the safety from not needing all of the `reinterpret_casts` outweighed the benefits from modularity.

Reviewed By: shayne-fletcher

Differential Revision: D31524302

fbshipit-source-id: 34c1c675a4fe41b9affdc8e75cc70a6c00122a50