fix bug in deferments with interrupted typechecking
Summary:
This fixes a but where scraping AST for class names didn't actually scrape all the classes.
This is fixed by running the full naming phase on the AST before scraping, because naming converts some `CIexpr (Id "MyClass")` nodes to `CI ("MyClass")`. The scraper is only able to scrape the second one.

Reviewed By: andrewjkennedy

Differential Revision: D31650988

fbshipit-source-id: 5561dfffa37f0f3178ab924aa1bb36cdd2bcdee2