Boundle map and allocator as a struct for the argument to the concurrent-hashmap write_map function closure
Summary:
I will be adding evictable allocators. As such, the `with` function that provides access to the shard's hashtable and allocator will now accept a closure with 3 arguments, of which 2 will be of the same "allocator" type: `fn(&HashMap, &MapAlloc, &MapAlloc)`.

This is very error prone. Since Rust has no named arguments, I introduce a struct called `Shard` that bundles these arguments and gives them a name!

```
/// This struct gives access to a shard, including its hashmap and its
/// allocators.
pub struct Shard<'shm, 'a, K, V, S> {
    pub map: &'a mut Map<'shm, K, V, S>,
    pub alloc_evictable: &'a MapAlloc<'shm>,
    pub alloc_non_evictable: &'a MapAlloc<'shm>,
}
```

Currently, it just contains a map and its allocator. In the future, there will be two allocators.

Differential Revision: D31109897

fbshipit-source-id: 9a78cda91cc100ec00e5bbc67829bfbeab6ddc62