Evictable allocators are fixed-size and only allocate one big chunk forever
Summary:
In a previous approach (D31196891), I put in code that would track large chunks as well, with the intention of free-ing them when evictable allocators are reset.

However, mjhostet noticed that actually, we should request a big chunk up front and never again allocate after that. As long as we don't touch a page, the memory won't be physically allocated anyways.

For evictable allocators, this works (because we know the maximum size up front), but for non-evictable allocators, it doesn't (theoretically they can grow indefinitely).

For evictable allocators, we should just refuse to allocate large objects (which would rid us of having to keep track of them). For non-evictable, for now, it's allocate-and-forget. I expect this to stay this way, as over time everything will become evictable.

-----

This diff introduces two configurable flags per allocator:

  1. is_fixed_size: Indicates the the allocator is of a fixed size. It will only allocate one chunk of that size. This is suitable for evictable allocators.

  2. chunk_size. The chunk size to use. For evictable allocators this will be MAX_GLOBAL_CACHE_SIZE/NUM_SHARDS (in the future). For non-evictable allocators, this will be a sensible constant value (e.g. 1 MiB)

If the allocator is of a fixed size (and thus evictable), it will refuse to allocate more underlying memory if it's current chunk is full. It will also refuse to allocate large objects (defined as 5% of the total allocator/chunk size). Clients will  have to make sure their request doesn't OOM the evictable allocator.

Reviewed By: mjhostet

Differential Revision: D31615616

fbshipit-source-id: a3d8333f11774a289d2eae19314d45e7b5fec27d