RFC: Disallow reading parent properties before `parent::__construct`
Summary:
# High Level Goal

As noted in the attached task, we (today) allow accessing inherited properties before calling `parent::__construct`:
```
abstract class MyParentClass {
  public function __construct(public string $str) {}
}
final class MySubClass extends MyParentClass {
  public function __construct(string $str) {
    PHP\var_dump($this->str);
    parent::__construct($str);
  }
}
```
This will dump `null`. This code, overall, is fairly nasty: it allows using private instance methods for initialization, which seems questionable. Let's error on this.

# Caveats / Concerns

The way I've gone about this is by grabbing the parent members that are expected to be initialized in their constructor (that is, initialized in `parent::__construct`) and used them as *part* of the initialization check. This required teaching legacy decls about properties that need initialization, something that gets dropped between translating a shallow class prop into a decl class prop.

Funny enough, this implementation only works at one layer for shallow decls, but works for `N` layers for legacy decls, as we've folded the entire hierarchy into the class.

Differential Revision: D30207030

fbshipit-source-id: 1597e4bceea3a8f10948756a2085a5a84619a01d