Add SDT flag to function types
Summary: We extend the idea of `<<__SupportDynamicType>>` on functions and methods to function types themselves. The idea is that an "SDT function type" represents functions that support dynamic calling, i.e. they have been type-checked under both their static types and assuming dynamic calling. For now, the attribute is stored internally, and isn't inspected except for pretty-printing. In future, we might support some syntax (e.g. `(dynamic function(C):D)`) and inspect the attribute on the function type instead of on methods and functions when type-checking calls.

Reviewed By: vsiles

Differential Revision: D31604823

fbshipit-source-id: 3dddf421f272514aa1bf676d2a30613113758031