Sync Iterable coeffects between HHVM and Hack
Summary:
https://fburl.com/scuba/php_errors/29trecx1 lists places where some methods which claim to be pure are calling methods which don't claim to be pure:

> Call to `LazyKVZipIterable::map()` requires `[defaults]` coeffects but `FeatureBank_BreakdownValuemap::__stringifyMapKeys()` provided `[]`

We need to somehow fix the code so that pure methods only call pure methods, all the way down.

Here's `LazyKVZipIterable::map()` in HHVM code:

https://www.internalfb.com/code/fbsource/[57c6890283801064053564901adbaebf93b57bf7]/fbcode/hphp/system/php/collections/collections.ns.php?lines=451-453

But this is `LazyKVZipIterable::map()` in the Hack typechecker's codebase:

https://www.internalfb.com/code/fbsource/[e5302970809bfb0bd6c4e62f4948cf9781101f16]/fbcode/hphp/hack/hhi/traits.hhi?lines=120

HHVM's version of the function signature is not as elaborate as the Hack typechecker's version. Let's sync the function signatures so HHVM sees the same coeffects.

Reviewed By: oulgen

Differential Revision: D31598078

fbshipit-source-id: fcfd04e167bbd7a6336fcae82b4768d293378a54