Remove some unnecessary gcc pinning
Summary:
These were both leftovers from T33723700, so remove `gcc` pinning, which
unbreaks using `-c fbcode.spit-dwarf=true` instead of `mode/opt-split-dwarf`
in some builds.

In the case of FBSerialize, the removal is essential a noop, as this was dir only
contained a library (and HHVM will to continue to build as gcc).

Fixes a build break from D31628607.

Reviewed By: alobbFB, luciang

Differential Revision: D31720704

fbshipit-source-id: 0307ec81f00409581c77c5eb1eae375d367af8cd