Invalidate shallow classes during shallow decl fanout
Summary:
During mass replays, I discovered a bug relating to shallow decl cache invalidation. We need to invalidate shallow classes during shallow decl fanout calculation since otherwise we'll have the wrong behavior in the rare event of a naming conflict.

In particular if during fanout calculation we fetch and cache the shallow decl of the losing file. That incorrect decl will persist over the span of typechecking. In the example in the test plan, what happens is the losing "IPAddress" class gets persisted, resulting in incorrect error calculations.

This diff fixes the problem by invalidating the shallow decl caches at the end of shallow decl fanout calculation. Note this would also be a problem in folded decl fanout, but it's not because we actually do invalidate shallow classes in do_redecl_phase2. However, with shallow fanout, we skip phase2, which is why we have this problem.

Differential Revision: D31722829

fbshipit-source-id: 9c7e4262686bb3f84096a97d00f67c6f8011e9dc