Don't redundantly override runtime options hhbbc::main()
Summary:
This function `hhbbc::main()` is difficult to read because we write several runtime options 3 or 4 times before we use them in `compile_repo()`. Tracing these is a pain, and this diff aims to make it more clear where the options change to make it faster/safer the next time someone wants to change an option.

Its still unclear to me why we did it this way before, but:
Currently:
1. `gd.load(false);` Loads global data from repo file and sets all relevant runtime options to these values
2. `RO::Load(ini, config);` Override runtime options with their default values in runtime-option.h
3. `hphp_process_init();` Override runtime options with values from global data as we did in step 1 (contains a call to gd.load())
4. lines 496-520: Override runtime options with values from global data as we just did in step 3.

What if instead we first set our runtime options to be their default values and then override them once with whatever is in the gd from our repo file?

----
Aside:
`RO::EvalEnforceGenericsUB ` is 1 or 2 in hhbbc when running in `repo-separate` mode
`RO::EvalEnforceGenericsUB` is 0 or 2 in hhbbc when running otherwise.

This seems wrong. History was hard to look at because things have been renamed. Going to fix this in diff on top of this one.
-----

Reviewed By: ricklavoie

Differential Revision: D31701294

fbshipit-source-id: 2deb12ce7cf39feb350dc1b79e38d87233a8fb39