Use -Drust-2018-idioms: stragglers (13/N)
Summary:
utils/arena_deserializer was easy now that everything else is migrated.

facebook/grpc/remote_execution has generated files, but I couldn't find
instructions to regenerate them. It looks like the generator tool is
`protobuf` which is version-locked to an old version 2.8.1. Rather than
fight with a tool upgrade, lets just opt-in to the legacy behavior:
`-Aelided-lifetimes-in-paths`.

Differential Revision: D31502334

fbshipit-source-id: c084a545a516b43fa087d099049260f33da5c372