Stop committing mcrouter codegen (#8918)
Summary:
Now that we're using thrift's build system instead of a custom one,
we have a working `thrift1`, so don't need to commit the codegen.

I'm not doing the usual submodule -> externalproject_add thing here as mcrouter
- does not use CMake
- does not publish regular releases

So, this is still going to be more work to update than most dependencies, but at
least we're no longer going to need to update the codegen by hand.

Pull Request resolved: https://github.com/facebook/hhvm/pull/8918

Test Plan: lego*, manually built on AWS (not on my mac as mcrouter does not build on MacOS)

Differential Revision: D31768051

fbshipit-source-id: 209be0b39d5b3a286d7c31c4791f3ba7264fa236