Don't silently eat exceptions from region selector
Summary:
We've always caught any exceptions thrown from the region
selector, and then silently returned an empty region. This means we'll
run the interpreter instead. It's not clear why we bothered to do
this, as the region selector shouldn't be throwing random
exceptions. The only exception that might be throw is FailedIRGen (a
punt).

The downside of this is that if something *does* throw an exception
incorrectly, we'll silently continue working and we might not be
aware. A common way this manifests is a misformatted FTRACE causing
the JIT to silently stop working.

This has bitten me more than once. Instead of silently eating the
exception, assert if we get anything other than FailedIRGen. That way
we'll know if something starts throwing.

Reviewed By: ottoni

Differential Revision: D31766599

fbshipit-source-id: 921c59714394072aa328770f497ec42b8b65b51c